Update to compile on Windows MinGW32
This should still compile with xcode on the Mac. I do not have a Mac to test though.
I updated the hid.c for mac and added hid_win for windows. Updated the hdiapi.h also to go with them.

Added a Makefile for mingw, not sure if the other makefile is used for mac so I left it alone.

converted usleep to not throw errors on mingw, should retain full compatibility with xcode on mac.