Fixed getArea1Sequence() and other decimal / hex corrections
getArea1Sequence() was getting decimal 24 when it should have been hex 0x24 or decimal 36. Since the documentation has everything in hex locations I switched most of code to be consistent. There is a few that are not making on this commit due to other changes I send in another commit.

Updated some of the info at the top about the locations of data